+++ internal_filesystem/lib/mpos/board/fri3d_2026.py
@@ -21,7 +21,7 @@ import drivers.display.st7789 as st7789
 
 import mpos.ui
 import mpos.ui.focus_direction
-from mpos import InputManager, IRManager
+from mpos import InputManager, IRManager, DeviceManager
 
 spi_bus = SPI.Bus(
     host=2,
@@ -218,6 +218,7 @@ mpos.ui.main_display.set_color_inversion(True)
 # touch pad interrupt TP Int is on ESP.IO13
 import drivers.indev.cst816s as cst816s
 i2c_bus = i2c.I2C.Bus(host=0, scl=18, sda=9, freq=400000, use_locks=False)
+DeviceManager.registerBus(i2c_bus=i2c_bus) # register because Time of Flight app needs it
 touch_dev = i2c.I2C.Device(bus=i2c_bus, dev_id=0x15, reg_bits=8)
 try:
     tindev=cst816s.CST816S(touch_dev,startup_rotation=lv.DISPLAY_ROTATION._180) # button in top left, good